requirements.txt
Django==1.7
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.14
django-celery==3.1.10
django-jsonfield==0.9.13
kombu==3.0.22
pytz==2014.7
requests==2.4.1
wsgiref==0.1.2
